upscale/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.1
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
prompt-toolkit==3.0.38
python-decouple==3.8
pytz==2023.3
redis==4.5.5
requests==2.30.0
singleton==0.1.0
six==1.16.0
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
